Analyse

Analyse

Usage

# S3 method for class 'character'
analyse(
  x,
  data,
  select_data = list(),
  nchains = getOption("mb.nchains", 3L),
  niters = getOption("mb.niters", 1000L),
  nthin = getOption("mb.nthin", 1L),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE),
  beep = getOption("mb.beep", TRUE),
  ...
)

Arguments

x

An object inheriting from class mb_model or a list of such objects.

data

The data frame to analyse.

select_data

A named list specifying the columns to select and their associated classes and values as well as transformations and scaling options.

nchains

A count of the number of chains.

niters

A count of the number of simulations to save per chain.

nthin

A count of the thining interval or NULL (in which case taken from model).

parallel

A flag indicating whether to perform the analysis in parallel if possible.

quiet

A flag indicating whether to disable tracing information.

glance

A flag indicating whether to print a model summary.

beep

A flag indicating whether to beep on completion of the analysis.

...

Additional arguments.